requirements.txt
aioredis==2.0.1
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
black==25.1.0
click>=8.0.0,<9.0.0
fastapi==0.115.13
flake8==7.3.0
graphql-core==3.2.6
greenlet==3.2.3
h11==0.16.0
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
isort==6.0.1
mccabe==0.7.0
mypy==1.16.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platform_common @ git+https://${PLATFORM_COMMON_TOKEN}@github.com/migoVanDingo/ed-platform-common@main
#-e /Users/bubz/Developer/web-dev/current-projects/ed-platform/ed-platform-common
platformdirs==4.3.8
pluggy==1.6.0
pycodestyle==2.14.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyflakes==3.4.0
Pygments==2.19.2
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
google-cloud-storage==3.6.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlmodel==0.0.24
starlette==0.46.2
strawberry-graphql==0.275.2
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
